Rework WebSocket handling

+++ src/microverse/__main__.py
@@ -12,7 +12,12 @@ from cyclopts import App
 app = App()
 
 @app.default
-def _main(*, environment: str = "environment", serve: bool = False):
+def _main(
+    *,
+    environment: str = "environment",
+    create_env: bool = True,
+    serve: bool = False,
+):
     environment = Path(environment)
     version = "0.1.4"
     here = Path(__file__).absolute().parent
@@ -25,17 +30,16 @@ def _main(*, environment: str = "environment", serve: bool = False):
     build_dir = Path("build").absolute()
     env_dir = Path("env").absolute()
     shutil.rmtree(build_dir, ignore_errors=True)
-    shutil.rmtree(env_dir, ignore_errors=True)
     build_dir.mkdir()
     
-    (Path(sys.prefix) / "share" / "empack" / "empack_config.yaml").write_text(empack_config)
+    if create_env:
+        shutil.rmtree(env_dir, ignore_errors=True)
+        call(f'micromamba create -f {environment / "environment.yml"} --platform emscripten-wasm32 --prefix {env_dir} --relocate-prefix "/" --yes')
 
-    def call(command: str):
-        subprocess.run(command, check=True, shell=True)
-
-    call(f'micromamba create -f {environment / "environment.yml"} --platform emscripten-wasm32 --prefix {env_dir} --relocate-prefix "/" --yes')
     for filename in (env_dir / "lib_js" / "pyjs").glob("*"):
         shutil.copy(filename, build_dir)
+
+    (Path(sys.prefix) / "share" / "empack" / "empack_config.yaml").write_text(empack_config)
     call(f"empack pack env --env-prefix {env_dir} --outdir {build_dir} --no-use-cache")
     call(f"empack pack dir --host-dir {environment / 'contents'} --mount-dir /contents --outname contents.tar.gz --outdir {build_dir}")
     call(f"empack pack append --env-meta {build_dir} --tarfile {build_dir / 'contents.tar.gz'}")
@@ -63,6 +67,10 @@ def _main(*, environment: str = "environment", serve: bool = False):
         print("Running server at http://127.0.0.1:8000")
         server = HTTPServer(("0.0.0.0", 8000), StaticHandler)
         server.serve_forever()
+
+
+def call(command: str):
+    subprocess.run(command, check=True, shell=True)
 
 
 def main():
